+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/contentshare/ContentShareController.swift
@@ -26,6 +26,21 @@ import Foundation
     /// - Parameter source: source of content to be shared
     func startContentShare(source: ContentShareSource)
 
+    /// Start sharing the content of a given `ContentShareSource`, with configurations.
+    ///
+    /// Once sharing has started successfully, `ContentShareObserver.contentShareDidStart` will
+    /// be notified. If sharing fails or stops, `ContentShareObserver.contentShareDidStop`
+    /// will be invoked with `ContentShareStatus` as the cause.
+    ///
+    /// This will call `VideoSource.addVideoSink(sink:)` on the provided source
+    /// and `VideoSource.removeVideoSink(sink:)` on the previously provided source.
+    ///
+    /// Calling this function repeatedly will replace the previous `ContentShareSource` as the one being transmitted.
+    ///
+    /// - Parameter source: source of content to be shared
+    /// - Parameter config: configurations of emitted video stream, e.g simulcast
+    func startContentShare(source: ContentShareSource, config: LocalVideoConfiguration)
+
     /// Stop sharing the content of a `ContentShareSource` that previously started.
     ///
     /// Once the sharing stops successfully, `ContentShareObserver.contentShareDidStop`