+++ AmazonChimeSDK/AmazonChimeSDKTests/audiovideo/DefaultAudioVideoControllerTests.swift
@@ -17,7 +17,6 @@ class DefaultAudioVideoControllerTests: CommonTestCase {
     var clientMetricsCollectorMock: ClientMetricsCollectorMock!
     var videoClientControllerMock: VideoClientControllerMock!
     var videoTileControllerMock: VideoTileControllerMock!
-    var defaultAudioClientMock: DefaultAudioClientMock!
     var defaultAudioVideoController: DefaultAudioVideoController!
 
     override func setUp() {
@@ -203,6 +202,28 @@ class DefaultAudioVideoControllerTests: CommonTestCase {
         XCTAssertNoThrow(try defaultAudioVideoController.startLocalVideo())
 
         verify(videoClientControllerMock.startLocalVideo()).wasCalled()
+    }
+
+    func testStartLocalVideoWithConfig() {
+        let config = LocalVideoConfiguration()
+        XCTAssertNoThrow(try defaultAudioVideoController.startLocalVideo(config: config))
+
+        verify(videoClientControllerMock.startLocalVideo(config: config)).wasCalled()
+    }
+
+    func testStartLocalVideoWithSource() {
+        let cameraCaptureSourceMock: CameraCaptureSourceMock = mock(CameraCaptureSource.self)
+        defaultAudioVideoController.startLocalVideo(source: cameraCaptureSourceMock)
+
+        verify(videoClientControllerMock.startLocalVideo(source: cameraCaptureSourceMock)).wasCalled()
+    }
+
+    func testStartLocalVideoWithSourceAndConfig() {
+        let config = LocalVideoConfiguration()
+        let cameraCaptureSourceMock: CameraCaptureSourceMock = mock(CameraCaptureSource.self)
+        defaultAudioVideoController.startLocalVideo(source: cameraCaptureSourceMock, config: config)
+
+        verify(videoClientControllerMock.startLocalVideo(source: cameraCaptureSourceMock, config: config)).wasCalled()
     }
 
     func testStopLocalVideo() {